Nat's comments
